Remove python 2.7 from the travis config.
There is an unknown issue in test_process_samples.py with writing
the kraken results to a temp file under python 2.7 that causes the parsing
method to fail. The same method is fine when run in test_parsing.py and
test_biom_create.py under python 2.7. So until I can figure out where the
problem is, I'm removing python 2.7 from the travis config.